+++ tests/sample/runner_property_test.py
@@ -1,0 +1,4 @@
+def test_runner(runner, sample):
+    """The runner property should return the runner associated with the sample"""
+
+    assert sample.runner['_id'] == runner['_id']